+++ lib/src/libset.h
@@ -110,7 +110,7 @@ int libset_set_int (const char *key, int val);
 
 int is_libset_var (const char *s);
 
-int libset_use_openmp (int n);
+int libset_use_openmp (double n);
 
 /* GUI setter functions */
 void set_xsect_hccme (const char *s);
